refactor: code structure refactor
