+++ src/mock/points.js
@@ -1,4 +1,4 @@
-import { getRandomArrayElement } from '../utils';
+import { getRandomArrayElement } from '../utils/utils.js';
 
 const mockPoints = [
   {
