--- src/utils.js
@@ -1,10 +1,0 @@
-import dayjs from 'dayjs';
-import {nanoid} from 'nanoid';
-
-export const getRandomArrayElement = (arr) => {
-  const item = arr[Math.floor(Math.random() * arr.length)];
-  return {...item, id: nanoid()};
-};
-export const transformData = (date, format) => date ? dayjs(date).format(format) : '';
-export const ucFirst = (str) => str[0]?.toUpperCase() + str?.slice(1);
-export const updateItem = (items, update) => items.map((item) => item.id === update.id ? update : item);
