+++ src/view/filter-view.js
@@ -1,6 +1,6 @@
 import { FilterType } from '../const.js';
 import AbstractView from '../framework/view/abstract-view.js';
-import { ucFirst } from '../utils.js';
+import { ucFirst } from '../utils/utils.js';
 const filtersNames = Object.values(FilterType);
 
 function createFilterTemplate(filters) {
