+++ src/view/sort-view.js
@@ -1,6 +1,6 @@
 import { SORTS } from '../const.js';
 import AbstractView from '../framework/view/abstract-view.js';
-import { ucFirst } from '../utils.js';
+import { ucFirst } from '../utils/utils.js';
 
 function createSortTemplate() {
   const renderBoardsSorts = () => SORTS.map((sort, index) => `
